% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{fast-save}
\alias{fast-save}
\alias{saveLZ4}
\alias{readLZ4}
\alias{saveZstd}
\alias{readZstd}
\alias{saveSnappy}
\alias{readSnappy}
\title{Alternative serialization Interface for Single Objects}
\usage{
saveLZ4(object, file, level = 1)

readLZ4(file)

saveZstd(object, file, level = -1)

readZstd(file)

saveSnappy(object, file)

readSnappy(file)
}
\arguments{
\item{object}{Object to serialize.}

\item{file}{A connection or the name of the file where the object is saved to or read from.}

\item{level}{Integer: the level of compression to be used.}
}
\description{
Functions to write a single \code{R} object to a compressed file and to restore it.
}
\details{
\code{lapply}, \code{sapply}, and \code{apply} are formula compatible versions of \code{base} \code{\link{lapply}}, \code{\link{sapply}} and \code{\link{apply}}
}
