% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{load.lbzip2.e}
\alias{load.lbzip2.e}
\title{Reload Datasets saved with lbzip2 and return it as a environment}
\usage{
load.lbzip2.e(...)
}
\arguments{
\item{...}{parameters for load.lbzip2()}
}
\value{
environment with the contents of the saved session
}
\description{
Reload Datasets saved with lbzip2 and return it as a environment
}
