% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{load.pigz}
\alias{load.pigz}
\title{Load R Objects saved with pigz compression}
\usage{
load.pigz(...)
}
\description{
This function is an alias for the internal load function
as files saved with pigz are compatible with it
}
