% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{preserve.state}
\alias{preserve.state}
\title{Save the session in the current working directory with a file name that
includes the time stamp and process id}
\usage{
preserve.state(prefix = "savepoint_", compression = "lbzip2", n.cores = 4)
}
\arguments{
\item{prefix}{the prefix to use,for the filename}

\item{compression}{type of file to save 'gz' or 'lbzip2', default: lbzip2}

\item{n.cores}{number of cores for save.image.fast or save.image.lbzip2, default = 4}
}
\value{
the name of the file that was saved in
}
\description{
Save the session in the current working directory with a file name that
includes the time stamp and process id
}
