% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{readRDS.lbzip2}
\alias{readRDS.lbzip2}
\title{Read a single R object in RDS format compressed with lbzip2}
\usage{
readRDS.lbzip2(file, refhook = NULL, n.cores = 4)
}
\arguments{
\item{file}{name of the file to save}

\item{refhook}{a hook function for handling reference objects}

\item{n.cores}{number of cores to use (default = 4)}
}
\description{
Read a single R object in RDS format compressed with lbzip2
}
