% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{readRDS.pigz}
\alias{readRDS.pigz}
\title{Save a single R object in RDS format that is compressed with pigz}
\usage{
readRDS.pigz(...)
}
\arguments{
\item{...}{parameters passed to readRDS}
}
\description{
this function delegates to the built in readRDS and is
provided for naming consistency
}
