% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{save.image.fast}
\alias{save.image.fast}
\title{Save the current workspace}
\usage{
save.image.fast(file = ".RData", n.cores = 4, safe = TRUE)
}
\arguments{
\item{file}{the name of the file where data will be saved}

\item{n.cores}{number of cores to use to compress}

\item{safe}{logical. If TRUE, a temporary file is used for creating the saved workspace.
The temporary file is renamed to file if the save succeeds. This preserves an existing
workspace file if the save fails, but at the cost of using extra disk space during the save.}
}
\description{
save.image.fast () is just a short-cut for ‘save my current workspace’,
i.e., save.fast(list = ls(all.names = TRUE), file = ".RData", envir = .GlobalEnv).
}
