% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{save.lbzip2}
\alias{save.lbzip2}
\title{Save R Object using lbzip2 parallel compression}
\usage{
save.lbzip2(..., list = character(),
  file = stop("'file' must be specified"), envir = parent.frame(),
  n.cores = 4, eval.promises = TRUE, precheck = TRUE)
}
\arguments{
\item{...}{the names of the objects to be saved (as symbols or character strings).}

\item{list}{A character vector containing the names of objects to be saved.}

\item{file}{the name of the file where data will be saved, the recommended suffix is .RDataFS}

\item{envir}{environment to search for objects to be saved.}

\item{n.cores}{number of cores to use to compress}

\item{eval.promises}{logical: should objects which are promises be forced before saving?}

\item{precheck}{logical: should the existence of the objects be checked before starting to
save (and in particular before opening the file/connection)?}
}
\description{
save writes an external representation of R objects to the specified file.
The objects can be read back from the file at a later date
by using the function load or attach (or data in some cases).
Files generated with this function can be loaded with load.lbzip2()
}
