% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds.R
\name{saveRDS.lbzip2}
\alias{saveRDS.lbzip2}
\title{Save a single R object in RDS format compressed with lbzip2}
\usage{
saveRDS.lbzip2(object, file = "", refhook = NULL, n.cores = 4)
}
\arguments{
\item{object}{R object to serialize}

\item{file}{name of the file to save}

\item{refhook}{a hook function for handling reference objects}

\item{n.cores}{number of cores to use (default = 4)}
}
\description{
Save a single R object in RDS format compressed with lbzip2
}
