% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_lbzip2.R
\name{load.lbzip2}
\alias{load.lbzip2}
\title{Reload Datasets saved with lbzip2}
\usage{
load.lbzip2(file, envir = parent.frame(), verbose = FALSE, n.cores = 4)
}
\arguments{
\item{envir}{the environment where the data should be loaded.}

\item{verbose}{should item names be printed during loading?}
}
\description{
Reload datasets written with the function save.lbzip2
}
