% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pigz.R
\name{load.pigz.e}
\alias{load.pigz.e}
\title{Load R Objects saved with pigz compression}
\usage{
load.pigz.e(...)
}
\arguments{
\item{...}{parameters for load}
}
\description{
This function is an alias for the internal load function
as files saved with pigz are compatible with it
}
