% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pigz.R
\name{load.pigz}
\alias{load.pigz}
\title{Load R Objects saved with pigz compression}
\usage{
load.pigz(file, envir = parent.frame(), verbose = FALSE)
}
\arguments{
\item{file}{the name of the file where data will be saved}

\item{envir}{the environment where the data will be loaded}

\item{verbose}{verbosity}
}
\description{
This function is an alias for the internal load function
as files saved with pigz are compatible with it
}
